requirements.txt
flask
Pillow
PyPDF2
